app/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cssselect2==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-extensions==3.1.5
django-jalali==6.0.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
easy-thumbnails==2.8.1
flake8==4.0.1
idna==3.3
jdatetime==4.0.0
kavenegar==1.1.2
lxml==4.8.0
mccabe==0.6.1
Pillow==9.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==22.0.0
pytz==2021.3
reportlab==3.6.8
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.3
